JsEvent: Make implementation internal & add xmldoc.
