+++ src/MudBlazor/Services/JsEvent/JsEventOptions.cs
@@ -2,8 +2,6 @@
 // MudBlazor licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System.Collections.Generic;
-
 namespace MudBlazor.Services
 {
     public class JsEventOptions
